refactor(backend-next): split app.ts into route modules
app.ts had reached 675 lines — 19% of the package in one file, holding all
nine routes inline alongside CORS and the OpenAPI document. It is now 108
lines and owns only composition; each resource's routes live in ./routes/,
mirroring @c15t/backend's layout.

That layout is not just tidiness. A single file holding every route grows past
what a reviewer can hold in mind, and it would have made the cutover diff —
where these routes replace the shipped ones — far harder to read than it needs
to be.

Route modules receive a RouteContext (app, options, run) rather than reaching
for module scope, so a route's dependencies are visible in its signature.

Largest file is now 376 lines. Behaviour-preserving: all 154 tests pass,
including the conformance suite, which caught a truncation mid-refactor when
an extraction dropped three subject routes.
